Moved MouseClickListener interface to event folder and refactor code
