+++ src/main/java/edu/kis/powp/jobs2d/events/MouseClickListener.java
@@ -1,4 +1,4 @@
-package edu.kis.powp.jobs2d.features;
+package edu.kis.powp.jobs2d.events;
 
 import java.awt.event.MouseEvent;
